# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for the owner-managed Project Hub catalog.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/sign-up:
    post:
      operationId: signUp
      tags: [authentication]
      summary: Create an account and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegistrationInput" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "400":
          description: Invalid registration details or owner setup code
        "409":
          description: Account or owner account already exists
        "503":
          description: Owner account setup is not configured securely
        "429":
          description: Too many registration attempts
  /auth/sign-in:
    post:
      operationId: signIn
      tags: [authentication]
      summary: Sign in and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "401":
          description: Email or password is incorrect
        "429":
          description: Too many sign-in attempts
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Get owner dashboard metrics
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Analytics overview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOverview" }
        "401":
          description: Authentication required
        "403":
          description: Owner access required
  /admin/projects:
    get:
      operationId: listAdminProjects
      tags: [admin]
      summary: List all projects including drafts
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Owner's projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
        "401":
          description: Authentication required
        "403":
          description: Owner access required
    post:
      operationId: createProject
      tags: [admin]
      summary: Create a project draft
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectInput" }
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "400":
          description: Invalid project details
        "401":
          description: Authentication required
        "403":
          description: Owner access required
  /admin/projects/{slug}:
    patch:
      operationId: updateProject
      tags: [admin]
      summary: Update a project
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Slug"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdate" }
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401":
          description: Authentication required
        "403":
          description: Owner access required
        "404":
          description: Project not found
    delete:
      operationId: archiveProject
      tags: [admin]
      summary: Archive a project without permanently deleting its history
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Slug"
      responses:
        "200":
          description: Project archived
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ActionResult" }
        "401":
          description: Authentication required
        "403":
          description: Owner access required
        "404":
          description: Project not found
  /admin/projects/{slug}/publication:
    put:
      operationId: setProjectPublication
      tags: [admin]
      summary: Publish or unpublish a project
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/Slug"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublicationInput" }
      responses:
        "200":
          description: Publication state updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        "401":
          description: Authentication required
        "403":
          description: Owner access required
        "404":
          description: Project not found
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: project_hub_session
  parameters:
    Slug:
      in: path
      name: slug
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
  schemas:
    AuthUser:
      type: object
      required: [id, email, displayName, role]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        displayName: { type: string }
        role: { type: string, enum: [user, admin] }
...
    AuthResult:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/AuthUser" }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, format: password, minLength: 12, maxLength: 128 }
    RegistrationInput:
      type: object
      required: [email, displayName, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        displayName: { type: string, minLength: 1, maxLength: 100 }
        password: { type: string, format: password, minLength: 12, maxLength: 128 }
        ownerSetupToken: { type: string, maxLength: 512 }
...
    Project:
      type: object
      required:
        - id
        - slug
        - title
        - shortDescription
        - description
        - category
        - tags
        - access
        - status
        - featured
        - isNew
        - version
        - views
        - downloads
        - favorites
        - releaseDate
        - updatedAt
        - screenshots
        - supportedDevices
        - requirements
        - whatsNew
        - changelog
        - fileAvailable
        - videoLocked
      properties:
        id: { type: string }
        slug: { type: string }
        title: { type: string }
        shortDescription: { type: string }
        description: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        access: { type: string, enum: [free, premium] }
        status: { type: string, enum: [draft, published, archived] }
        featured: { type: boolean }
        isNew: { type: boolean }
        thumbnailUrl: { type: ["string", "null"] }
        previewVideoUrl: { type: ["string", "null"] }
        version: { type: string }
        fileSize: { type: ["string", "null"] }
        views: { type: integer }
        downloads: { type: integer }
        favorites: { type: integer }
        releaseDate: { type: string, format: date }
        updatedAt: { type: string, format: date-time }
        screenshots:
          type: array
          items: { type: string }
        supportedDevices:
          type: array
          items: { type: string }
        requirements:
          type: array
          items: { type: string }
        whatsNew: { type: string }
        changelog:
          type: array
          items: { type: string }
        fileAvailable: { type: boolean }
        videoLocked: { type: boolean }
...
    ProjectInput:
      type: object
      required: [slug, title, shortDescription, description, category, access]
      properties:
        slug: { type: string, minLength: 2, maxLength: 100 }
        title: { type: string, minLength: 1, maxLength: 120 }
        shortDescription: { type: string, minLength: 1, maxLength: 220 }
        description: { type: string, minLength: 1, maxLength: 10000 }
        category: { type: string, minLength: 1, maxLength: 60 }
        tags:
          type: array
          items: { type: string, maxLength: 40 }
        access: { type: string, enum: [free, premium] }
        featured: { type: boolean }
        isNew: { type: boolean }
        thumbnailUrl: { type: ["string", "null"] }
        previewVideoUrl: { type: ["string", "null"] }
        version: { type: string, maxLength: 40 }
        fileSize: { type: ["string", "null"] }
        releaseDate: { type: string, format: date }
        screenshots:
          type: array
          items: { type: string }
        supportedDevices:
          type: array
          items: { type: string }
        requirements:
          type: array
          items: { type: string }
        whatsNew: { type: string }
        changelog:
          type: array
          items: { type: string }
    ProjectUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        shortDescription: { type: string, minLength: 1, maxLength: 220 }
        description: { type: string, minLength: 1, maxLength: 10000 }
        category: { type: string, minLength: 1, maxLength: 60 }
        tags:
          type: array
          items: { type: string, maxLength: 40 }
        access: { type: string, enum: [free, premium] }
        featured: { type: boolean }
        isNew: { type: boolean }
        thumbnailUrl: { type: ["string", "null"] }
        previewVideoUrl: { type: ["string", "null"] }
        version: { type: string, maxLength: 40 }
        fileSize: { type: ["string", "null"] }
        releaseDate: { type: string, format: date }
        screenshots:
          type: array
          items: { type: string }
        supportedDevices:
          type: array
          items: { type: string }
        requirements:
          type: array
          items: { type: string }
        whatsNew: { type: string }
        changelog:
          type: array
          items: { type: string }
    PublicationInput:
      type: object
      required: [published]
      properties:
        published: { type: boolean }
...
    ActionResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: ["string", "null"] }
...
    AdminOverview:
      type: object
      required: [projects, published, views, downloads, favorites, users, subscriptions, popular]
      properties:
        projects: { type: integer }
        published: { type: integer }
        views: { type: integer }
        downloads: { type: integer }
        favorites: { type: integer }
        users: { type: integer }
        subscriptions: { type: integer }
        popular:
          type: array
          items: { $ref: "#/components/schemas/Project" }