# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the owner-managed Project Hub catalog.
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthResult {
  user: AuthUser;
}