# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the owner-managed Project Hub catalog.
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserRole } from './authUserRole';

export interface AuthUser {
  id: string;
  email: string;
  displayName: string;
  role: AuthUserRole;
}