# lib/api-zod/src/generated/types/loginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the owner-managed Project Hub catalog.
 * OpenAPI spec version: 0.1.0
 */

export interface LoginInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 12
     * @maxLength 128
     */
  password: string;
}