# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for the owner-managed Project Hub catalog.
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  displayName: string;
  /**
     * @minLength 12
     * @maxLength 128
     */
  password: string;
  /** @maxLength 512 */
  ownerSetupToken?: string;
}